Add blackjack rules and hide them
This feature added the rules of blackjack with AI generation. The display: none; class was added to hide them, so that in future development JavaScript can be used to toggle this class on or off, to display the rules when the button is clicked. important had to be used to activate the class.